This change aggregates a couple of improvements for the perf client and server.
Some changes make the client and server more efficient.

Two other commits add stats to the client.

Output is now created in 2s intervals and looks like:

```
Overall stats:
RPS: 0 (2 requests in 14.01s)
Success rate: 100.00%

Stream metrics:

      │ Duration  │ FBL       | Upload Throughput | Download Throughput
──────┼───────────┼───────────┼───────────────────┼────────────────────
 AVG  │     4.96s │    0.00ns │      404.00 MiB/s │        402.62 MiB/s
 P0   │     4.94s │    0.00ns │      400.50 MiB/s │        397.00 MiB/s
 P10  │     4.95s │    0.00ns │      400.75 MiB/s │        397.25 MiB/s
 P50  │     4.95s │    0.00ns │      400.75 MiB/s │        397.25 MiB/s
 P90  │     4.97s │    1.00ms │      407.50 MiB/s │        408.25 MiB/s
 P100 │     4.97s │    1.00ms │      407.50 MiB/s │        408.25 MiB/s
```